+++ playwright-e2e/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./config.utils";
 export * from "./utils.fixtures";
 export * from "./cleardown.utils";
 export * from "./reporting.utils";
+export * from "./newHearingSessionUIValidation.utils";
